zsh: Fix a security issue. Wow.
Commit b31ccb1e2813014c3039aad259f3b4d73d02105a, from 2014/01/30, added
some dirs under .local to LD_LIBRARY_PATH, to make .local installs of
libraries more usable. Turns out ld.so has the bizarre behavior that an
empty element, such as you get when $LD_LIBRARY_PATH is empty and you
set it to "/something:$LD_LIBRARY_PATH", is treated as meaning the
current directory.

Noticed this while investigating extreme slowness when using sshfs over a
connection with 250 ms of organic latency. Thought it was an issue with
sshfs serializing stat calls or something, turns out it was ld.so stat'ing
nonexistent libraries in . over and over.